feat(02-03): integration tests for profile + transponder controllers
- RacerProfileControllerIT: 8 tests covering GET/PATCH profile, email/role field protection, membership CRUD + duplicate 409
- TransponderControllerIT: 6 tests covering CRUD, cross-user isolation, system-wide duplicate 409, anonymous 401

Auto-fixed issues (Rule 1/2):
- CarTagCategory.getSortOrder() returned int (widened) causing Hibernate schema validation to expect INTEGER instead of SMALLINT — fixed getter/setter to use short type; updated CarTagCategoryService casts accordingly
- SecurityConfig lacked AuthenticationEntryPoint so anonymous requests returned 403 instead of 401 — added HttpStatusEntryPoint(UNAUTHORIZED) to exceptionHandling

+++ app/src/main/java/dev/monkeypatch/rctiming/domain/car/CarTagCategoryService.java
@@ -34,7 +34,7 @@ public class CarTagCategoryService {
     public CarTagCategoryDto create(CreateCarTagCategoryRequest request) {
         CarTagCategory category = new CarTagCategory();
         category.setName(request.name());
-        category.setSortOrder(request.sortOrder() != null ? request.sortOrder() : 0);
+        category.setSortOrder(request.sortOrder() != null ? request.sortOrder().shortValue() : (short) 0);
         category.setCreatedAt(Instant.now());
         return CarTagCategoryDto.from(carTagCategoryRepository.save(category));
     }
@@ -43,7 +43,7 @@ public class CarTagCategoryService {
         CarTagCategory category = getCategoryOrThrow(id);
         category.setName(request.name());
         if (request.sortOrder() != null) {
-            category.setSortOrder(request.sortOrder());
+            category.setSortOrder(request.sortOrder().shortValue());
         }
         return CarTagCategoryDto.from(carTagCategoryRepository.save(category));
     }
